move generic files to package
